refactor(story): reorganize into phases/projects structure with queryOptions pattern
- Move StoryPhase/ → phases/story-phase/ with story-phase.queries.ts
- Move Projects/ → projects/ (lowercase) with projects.queries.ts
- Replace separate key factory objects with co-located queryOptions factories
- Inline thin hooks (useStoryDetail, useStoryHistory, useProjectHome, useProjectList)
- Consolidate duplicated endpoint constants into shared/story.constants.ts
- Move cross-cutting domain types into shared/story.types.ts
- Rename events/baseEvents.ts → base-events.ts, eventRouter.ts → event-router.ts
- Mutations now invalidate via queryOptions(...).queryKey instead of key objects
