+++ frontend/src/features/story/projects/projects.types.ts
@@ -1,0 +1,7 @@
+export type StoryResponseType = {
+  id: string;
+  projectId: string;
+  storyText: string;
+  userInputText: string;
+  meta: Record<string, unknown>;
+};
